refactor: add underscores to mark protected (functions, modules)
