+++ mostlyai/client/api.py
@@ -26,8 +26,8 @@ from mostlyai.client.synthetic_datasets import (
     _MostlySyntheticDatasetsClient,
     _MostlySyntheticProbesClient,
 )
-from mostlyai.client.base_utils import _convert_to_base64
-from mostlyai.client.mostly_utils import (
+from mostlyai.client._base_utils import _convert_to_base64
+from mostlyai.client._mostly_utils import (
     ShareableResource,
     _read_table_from_path,
     _harmonize_sd_config,
