+++ tests/unit/mostlyai/test_naming_conventions.py
@@ -1,10 +1,10 @@
 import pytest
 from datamodel_code_generator.reference import camel_to_snake
 
-from mostlyai.client.naming_conventions import (
-    snake_to_camel,
-    map_snake_to_camel_case,
-    map_camel_to_snake_case,
+from mostlyai.client._naming_conventions import (
+    _snake_to_camel,
+    _map_snake_to_camel_case,
+    _map_camel_to_snake_case,
 )
 
 
@@ -17,7 +17,7 @@ from mostlyai.client.naming_conventions import (
     ],
 )
 def test__snake_to_camel_param(snake_str, expected):
-    assert snake_to_camel(snake_str) == expected
+    assert _snake_to_camel(snake_str) == expected
 
 
 def test_map_snake_to_camel_case():
@@ -55,7 +55,7 @@ def test_map_snake_to_camel_case():
         "emptyStringKey": "",
     }
 
-    assert map_snake_to_camel_case(input_data) == expected_output
+    assert _map_snake_to_camel_case(input_data) == expected_output
 
 
 @pytest.mark.parametrize(
@@ -108,4 +108,4 @@ def test_map_camel_to_snake_case():
         "empty_string_key": "",
     }
 
-    assert map_camel_to_snake_case(input_data) == expected_output
+    assert _map_camel_to_snake_case(input_data) == expected_output
