+++ tests/unit/mostlyai/test_utils.py
@@ -23,11 +23,11 @@ from mostlyai.client.model import (
     SyntheticTableConfig,
     SyntheticProbeConfig,
 )
-from mostlyai.client.base_utils import (
+from mostlyai.client._base_utils import (
     _convert_to_base64,
     _read_table_from_path,
 )
-from mostlyai.client.mostly_utils import (
+from mostlyai.client._mostly_utils import (
     _job_wait,
     _harmonize_sd_config,
 )
